requirements.txt
beautifulsoup4~=4.12.2
chromadb~=0.4.22
langchain~=0.1.0
langchain-community
langchain-openai
python-dotenv~=1.0.0
tiktoken~=0.5.2
